setup restricted routes and temporary nav
