--- client/src/routes/Pages/Auth/index.js
@@ -1,17 +1,0 @@
-import * as React from 'react'
-
-import { Routes, BrowserRouter as Router, Route } from "react-router-dom";
-
-import Login from './Login'
-import Register from './Register'
-
-const Auth = () => {
-    return (
-        <Routes>
-            <Route path="/login" element={<Login/>} />
-            <Route path="/register" element={<Register/>} />
-        </Routes>
-    )
-}
-
-export default Auth
